requirements.txt
Django
psycopg2-binary
python-dotenv
pandas
openpyxl
reportlab
boto3
django-storages
djangorestframework
django-extensions
pygraphviz
django-cors-headers
channels
channels-redis
paho-mqtt
daphne
